fix: protect api routing

+++ src/pages/api/cards/[id].js
@@ -1,11 +1,14 @@
 import Item from "../../../ions/models/item.model";
 import dbConnect from "../../../ions/database/index";
+import { getSession } from "next-auth/react";
 
 const handler = async (request, response) => {
 	const {
 		query: { id },
 		method,
 	} = request;
+	const session = await getSession({ req: request });
+	console.log(session);
 
 	await dbConnect();
 
@@ -38,15 +41,18 @@ const handler = async (request, response) => {
 			break;
 
 		case "DELETE" /* Delete a model by its ID */:
-			try {
-				const deletedItem = await Item.deleteOne({ _id: id });
-				if (!deletedItem) {
-					return response.status(400).json({ success: false });
+			if (!session) {
+				response.status(401).send("Not authorized");
+			} else
+				try {
+					const deletedItem = await Item.deleteOne({ _id: id });
+					if (!deletedItem) {
+						return response.status(400).json({ success: false });
+					}
+					response.status(200).json({ success: true, data: {} });
+				} catch (error) {
+					response.status(400).json({ success: false });
 				}
-				response.status(200).json({ success: true, data: {} });
-			} catch (error) {
-				response.status(400).json({ success: false });
-			}
 			break;
 
 		default:
